feat: float64 implementation; verified k51n126v8

+++ python/setup.py
@@ -157,8 +157,8 @@ def setup_A_Wbit_D(path):
     # output
     out0 = np.mod(np.matmul(input_0, np.transpose(np.vstack((poly1, poly2)))), 2)
     out1 = np.mod(np.matmul(input_1, np.transpose(np.vstack((poly1, poly2)))), 2)
-    Wout0 = np.sum(out0, axis=1, dtype=np.int32)
-    Wout1 = np.sum(out1, axis=1, dtype=np.int32)
+    Wout0 = np.sum(out0, axis=1, dtype=np.uint8)
+    Wout1 = np.sum(out1, axis=1, dtype=np.uint8)
     zidx0 = (Wout0 == 0).reshape(-1, 1)
     zidx1 = (Wout1 == 0).reshape(-1, 1)
     W_weight = np.stack((Wout0, Wout1), axis=0).T.astype(np.uint8)
@@ -180,17 +180,17 @@ def setup_A_Wbit_D(path):
     basis = np.arange(0, num_total_states, 2 ** code_config["bch_config"]["M"])
     As = []
     for valid_starting_state in basis:
-        A = np.zeros(shape=(num_total_states, 1), dtype=np.uint64)
+        A = np.zeros(shape=(num_total_states, 1), dtype=np.float64)
         A[valid_starting_state] = 1
         As.append(A)
 
     # set up W
     # W_in: [input] x [num_states] x [max weight for one meta-stage]
-    W = np.stack((Wcoef0, Wcoef1), axis=0).astype(np.uint64)  # horizontal stack
+    W = np.stack((Wcoef0, Wcoef1), axis=0).astype(np.uint8)  # horizontal stack
 
     # set up D
     # D_in: [num_states] x [input]
-    D = np.stack((dst_0, dst_1), axis=1).astype(np.uint64)
+    D = np.stack((dst_0, dst_1), axis=1).astype(np.uint32)
 
     return As, W_weight, D, basis, num_trellis_stages
